+++ package.json
@@ -31,6 +31,7 @@
     "papaparse": "^5.3.1",
     "parquetjs-lite": "^0.8.7",
     "pg": "^8.6.0",
+    "preview": "https://github.com/multiprocessio/preview",
     "react": "^17.0.2",
     "react-ace": "^9.4.1",
     "react-dom": "^17.0.2",
