+++ ui/ProgramPanel.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { MODE, RPC } from '../shared/constants';
 import { LANGUAGES, SupportedLanguages } from '../shared/languages';
+import { Shape } from '../shared/shape';
 import { PanelResult, ProgramPanelInfo } from '../shared/state';
 import { asyncRPC } from './asyncRPC';
 import { Select } from './component-library/Select';
@@ -9,14 +10,14 @@ export function evalProgramPanel(
   panel: ProgramPanelInfo,
   panelResults: Array<PanelResult>,
   indexIdMap: Record<number, string>
-): Promise<{ value: any; preview: string; stdout: string }> {
+): Promise<{ value: any; preview: string; stdout: string; shape: Shape }> {
   const program = panel.program;
 
   if (MODE === 'desktop') {
     return asyncRPC<
       ProgramPanelInfo & { indexIdMap: Record<number, string> },
       null,
-      { value: any; preview: string; stdout: string }
+      { value: any; preview: string; stdout: string; shape: Shape }
     >(RPC.EVAL_PROGRAM, null, {
       ...panel,
       indexIdMap,
